deps: update V8 to 8.7.220
PR-URL: https://github.com/nodejs/node/pull/35700

+++ deps/v8/src/builtins/builtins-intl.cc
@@ -1034,8 +1034,8 @@ BUILTIN(SegmentsPrototypeContaining) {
                                      Object::ToInteger(isolate, index));
   double const n = index->Number();
 
-  RETURN_RESULT_OR_FAILURE(
-      isolate, JSSegments::Containing(isolate, segments, static_cast<int>(n)));
+  RETURN_RESULT_OR_FAILURE(isolate,
+                           JSSegments::Containing(isolate, segments, n));
 }
 
 // ecma402 #sec-%segmentsprototype%-@@iterator
